requirements.txt
fastapi>=0.115.11
langchain-openai>=0.3.7
langgraph>=0.3.2
python-dotenv>=1.0.1
uvicorn>=0.34.0